rename plugin module
