+++ services/web/server/src/simcore_service_webserver/login/handlers_registration.py
@@ -12,7 +12,7 @@ from servicelib.mimetype_constants import MIMETYPE_APPLICATION_JSON
 
 from .._meta import API_VTAG
 from ..groups_api import auto_add_user_to_groups, auto_add_user_to_product_group
-from ..invitations import is_service_invitation_code
+from ..invitations_plugin import is_service_invitation_code
 from ..products.plugin import Product, get_current_product
 from ..security.api import encrypt_password
 from ..session_access import on_success_grant_session_access_to, session_access_required
