+++ services/web/server/src/simcore_service_webserver/login/plugin.py
@@ -19,7 +19,7 @@ from ..db import setup_db
 from ..db_settings import get_plugin_settings as get_db_plugin_settings
 from ..email import setup_email
 from ..email_settings import get_plugin_settings as get_email_plugin_settings
-from ..invitations import setup_invitations
+from ..invitations_plugin import setup_invitations
 from ..products.plugin import ProductName, list_products, setup_products
 from ..redis import setup_redis
 from ..rest import setup_rest
