+++ services/web/server/tests/unit/with_dbs/01/studies_dispatcher/test_studies_dispatcher_projects.py
@@ -91,6 +91,8 @@ async def test_add_new_project_from_model_instance(
     file_picker_id: NodeID,
     viewer_id: NodeID,
 ):
+    assert client.app
+
     mock_directorv2_api = mocker.patch(
         "simcore_service_webserver.director_v2_api.create_or_update_pipeline",
         return_value=None,
